fix(picom): disable shadows for menus
Disable shadows for menus to fix thick borders appearing around menus
in Google Chrome. This issue is present when enabling Google's new
Material You redesign of Chrome in
[flags](chrome://flags/#chrome-refresh-2023)

Note to self: it looks like the redesign has hard-coded shadows.
Check if Google will roll out an update that allows shadows to
be implemented by the user's compositor and reenable shadows for menus
in Picom's config.