fix(zsh): display alternate prompt when using a tty
use an alternate prompt while in a tty because my prompt uses a unicode
symbol(➜) which does not display correctly while in a non graphical
environment. in its place is a "missing element" character which is
unaesthetic. in the tty, the unicode symbol is now replaced with a
character based element; an arrow(->)